Increase specificity from 1-0-0-0-0 to 1-0-20-0-0
This fixes the hyphenation still occurring on some websites like the
stackexchange.com network. Those sites combine element and class
selectors with !important to attain a specificity of 1-0-0-1-1, which
overrode our previous specificity of 1-0-0-0-0.

By adding 10 copies of :not(#a#b), the ID specificity is raised to 20.
Each :not(...) has the specificity of its children. #a#b is of
specificity 2-0-0, but it's impossible for an element to have two IDs.
The :not of "impossible" is "always", so this works for all elements
no matter the ID, and selects the same elements as "*".